First Real Commit
Have player movement. Player moves with WASD, looks towards where they are
going. They only turn so fast so they take a second to look where they are
going. The player can lunge forward in the direction they are facing. This
lunge is what the player uses to tag the hiders (AI controlled agents).

AI movement is controlled by Unity's Nav Mesh Agents. The AI paths towards
pre-set hiding places. If they can see the player (with a raycast), they
will move to another random hiding spot.

The game director spawns a number of AI hiders for the player to go catch.
It sets them up, shows them the player and the hiding spots and then has
an action that triggers when the player catches a hider.

Camera - the camera follows the player around. It lags behind the player
and slowly catches up.